Add trotter.py file
