+++ src/qibo/core/adiabatic.py
@@ -1,5 +1,6 @@
 from qibo.config import raise_error
 from qibo.core import hamiltonians
+from qibo.core.trotter import TrotterCircuit
 
 
 class AdiabaticHamiltonian:
@@ -63,6 +64,6 @@ class SymbolicAdiabaticHamiltonian(BaseAdiabaticHamiltonian):
         if accelerators is not None:
             self.accelerators = accelerators
             self.memory_device = memory_device
-        self.trotter_circuit = hamiltonians.SymbolicHamiltonian.TrotterCircuit(
+        self.trotter_circuit = TrotterCircuit(
             ham.nqubits, ham.terms, dt, self.accelerators, self.memory_device)
         return self.trotter_circuit.circuit
