+++ src/qibo/core/hamiltonians.py
@@ -382,28 +382,11 @@ class SymbolicHamiltonian(hamiltonians.SymbolicHamiltonian):
         raise_error(NotImplementedError, "Hamiltonian matmul to {} not "
                                          "implemented.".format(type(o)))
 
-    class TrotterCircuit:
-
-        def __init__(self, nqubits, terms, dt, accelerators, memory_device):
-            from qibo.models import Circuit
-            self.gates = {}
-            self.dt = dt
-            self.circuit = Circuit(nqubits, accelerators=accelerators,
-                                   memory_device=memory_device)
-            for term in itertools.chain(terms, terms[::-1]):
-                gate = term.expgate(dt / 2.0)
-                self.gates[gate] = term
-                self.circuit.add(gate)
-
-        def set_dt(self, dt):
-            params = {gate: term.exp(dt / 2.0) for gate, term in self.gates.items()}
-            self.dt = dt
-            self.circuit.set_parameters(params)
-
     def circuit(self, dt, accelerators=None, memory_device="/CPU:0"):
         if self.trotter_circuit is None:
-            self.trotter_circuit = self.TrotterCircuit(self.nqubits, self.terms, dt,
-                                                       accelerators, memory_device)
+            from qibo.core.trotter import TrotterCircuit
+            self.trotter_circuit = TrotterCircuit(self.nqubits, self.terms, dt,
+                                                  accelerators, memory_device)
         elif dt != self.trotter_circuit.dt:
             self.trotter_circuit.set_dt(dt)
         return self.trotter_circuit.circuit
